docs: seperate example site wrapper into seperate components
