+++ example/src/SiteWrapper.tsx
@@ -1,16 +1,9 @@
-import * as React from "react";
-import { NavLink as ReactRouterNavLink, withRouter } from "react-router-dom";
+import React from "react";
 
-import {
-  Site,
-  Nav,
-  Grid,
-  List,
-  Button,
-  Form,
-  RouterContextProvider,
-  NavItemProps,
-} from "tabler-react";
+import { Site } from "tabler-react";
+import SiteHeader from "./SiteHeader";
+import SiteFooter from "./SiteFooter";
+import SiteNav from "./SiteNav";
 
 type Props = {
   children: React.ReactNode;
@@ -20,336 +13,14 @@ type State = {
   notificationsObjects: any[];
 };
 
-const navBarItems: Array<NavItemProps> = [
-  {
-    value: "Home",
-    to: "/",
-    icon: "home",
-    linkAs: ReactRouterNavLink,
-    linkProps: { exact: true },
-  },
-  {
-    value: "Interface",
-    icon: "box",
-    href: "#interface",
-    subItemsObjects: [
-      {
-        value: "Cards Design",
-        to: "/cards",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Charts",
-        to: "/charts",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Pricing Cards",
-        to: "/pricing-cards",
-        linkAs: ReactRouterNavLink,
-      },
-    ],
-  },
-  {
-    value: "Components",
-    icon: "calendar",
-    href: "#components",
-    subItemsObjects: [
-      {
-        value: "Maps",
-        to: "/maps",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Icons",
-        to: "/icons",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Store",
-        to: "/store",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Blog",
-        to: "/blog",
-        linkAs: ReactRouterNavLink,
-      },
-    ],
-  },
-  {
-    value: "Pages",
-    icon: "file",
-    href: "#pages",
-    subItemsObjects: [
-      {
-        value: "Profile",
-        to: "/profile",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Login",
-        to: "/login",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Register",
-        to: "/register",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Forgot password",
-        to: "/forgot-password",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "400 error",
-        to: "/400",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "401 error",
-        to: "/401",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "403 error",
-        to: "/403",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "404 error",
-        to: "/404",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "500 error",
-        to: "/500",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "503 error",
-        to: "/503",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Email",
-        to: "/email",
-        linkAs: ReactRouterNavLink,
-      },
-      {
-        value: "Empty page",
-        to: "/empty-page",
-        linkAs: ReactRouterNavLink,
-      },
-      { value: "RTL", to: "/rtl", linkAs: ReactRouterNavLink },
-    ],
-  },
-  {
-    value: "Forms",
-    to: "/form-elements",
-    icon: "check-square",
-    linkAs: ReactRouterNavLink,
-  },
-  {
-    value: "Gallery",
-    to: "/gallery",
-    icon: "image",
-    linkAs: ReactRouterNavLink,
-  },
-  {
-    icon: "file-text",
-    value: "Documentation",
-    to:
-      process.env.NODE_ENV === "production"
-        ? "https://tabler.github.io/tabler-react/documentation"
-        : "/documentation",
-  },
-];
-
-const accountDropdownProps = {
-  avatarURL: "./demo/faces/female/25.jpg",
-  name: "Jane Pearson",
-  description: "Administrator",
-  options: [
-    { icon: "user", value: "Profile" },
-    { icon: "settings", value: "Settings" },
-    { icon: "mail", value: "Inbox", badge: "6" },
-    { icon: "send", value: "Message" },
-    { isDivider: true },
-    { icon: "help-circle", value: "Need help?" },
-    { icon: "log-out", value: "Sign out" },
-  ],
-};
-
 class SiteWrapper extends React.Component<Props, State> {
-  state = {
-    notificationsObjects: [
-      {
-        unread: true,
-        avatarURL: "demo/faces/male/41.jpg",
-        message: (
-          <React.Fragment>
-            <strong>Nathan</strong> pushed new commit: Fix page load performance
-            issue.
-          </React.Fragment>
-        ),
-        time: "10 minutes ago",
-      },
-      {
-        unread: true,
-        avatarURL: "demo/faces/female/1.jpg",
-        message: (
-          <React.Fragment>
-            <strong>Alice</strong> started new task: Tabler UI design.
-          </React.Fragment>
-        ),
-        time: "1 hour ago",
-      },
-      {
-        unread: false,
-        avatarURL: "demo/faces/female/18.jpg",
-        message: (
-          <React.Fragment>
-            <strong>Rose</strong> deployed new version of NodeJS REST Api // V3
-          </React.Fragment>
-        ),
-        time: "2 hours ago",
-      },
-    ],
-  };
-
   render() {
-    const notificationsObjects = this.state.notificationsObjects || [];
-    const unreadCount = this.state.notificationsObjects.reduce(
-      (a, v) => a || v.unread,
-      false
-    );
     return (
       <Site.Wrapper
-        headerProps={{
-          href: "/",
-          alt: "Tabler React",
-          imageURL: "./demo/brand/tabler.svg",
-          navItems: (
-            <Nav.Item link={false} className="d-none d-md-flex">
-              <Button
-                href="https://github.com/tabler/tabler-react"
-                target="_blank"
-                outline
-                size="sm"
-                RootComponent="a"
-                color="primary"
-              >
-                Source code
-              </Button>
-            </Nav.Item>
-          ),
-          notificationsTray: {
-            notificationsObjects,
-            markAllAsRead: () =>
-              this.setState(
-                () => ({
-                  notificationsObjects: this.state.notificationsObjects.map(
-                    v => ({ ...v, unread: false })
-                  ),
-                }),
-                () =>
-                  setTimeout(
-                    () =>
-                      this.setState({
-                        notificationsObjects: this.state.notificationsObjects.map(
-                          v => ({ ...v, unread: true })
-                        ),
-                      }),
-                    5000
-                  )
-              ),
-            unread: unreadCount,
-          },
-          accountDropdown: accountDropdownProps,
-          searchBar: (
-            <Form.Input
-              icon="search"
-              position="prepend"
-              placeholder="Search"
-              tabIndex={-1}
-              light
-            />
-          ),
-        }}
-        navProps={{ itemsObjects: navBarItems }}
-        routerContextComponentType={withRouter(RouterContextProvider as any)}
-        footerProps={{
-          links: [
-            <a href="#1" key="1">
-              First Link
-            </a>,
-            <a href="#2" key="2">
-              Second Link
-            </a>,
-            <a href="#3" key="3">
-              Third Link
-            </a>,
-            <a href="#4" key="4">
-              Fourth Link
-            </a>,
-            <a href="#5" key="5">
-              Five Link
-            </a>,
-            <a href="#6" key="6">
-              Sixth Link
-            </a>,
-            <a href="#7" key="7">
-              Seventh Link
-            </a>,
-            <a href="#8" key="8">
-              Eigth Link
-            </a>,
-          ],
-          note:
-            "Premium and Open Source dashboard template with responsive and high quality UI. For Free!",
-          copyright: (
-            <React.Fragment>
-              Copyright © 2019
-              <a href="."> Tabler-react</a>. Theme by
-              <a
-                href="https://codecalm.net"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                {" "}
-                codecalm.net
-              </a>{" "}
-              All rights reserved.
-            </React.Fragment>
-          ),
-          nav: (
-            <React.Fragment>
-              <Grid.Col auto={true}>
-                <List className="list-inline list-inline-dots mb-0">
-                  <List.Item className="list-inline-item">
-                    <a href="./docs/index.html">Documentation</a>
-                  </List.Item>
-                  <List.Item className="list-inline-item">
-                    <a href="./faq.html">FAQ</a>
-                  </List.Item>
-                </List>
-              </Grid.Col>
-              <Grid.Col auto={true}>
-                <Button
-                  href="https://github.com/tabler/tabler-react"
-                  size="sm"
-                  outline
-                  color="primary"
-                  RootComponent="a"
-                >
-                  Source code
-                </Button>
-              </Grid.Col>
-            </React.Fragment>
-          ),
-        }}
+        header={SiteHeader}
+        nav={SiteNav}
+        navIsSide={true}
+        footer={SiteFooter}
       >
         {this.props.children}
       </Site.Wrapper>
